ILIAR/pyproject.toml
[project]
name = "torchtmpl"
version = "0.0.1"
description = "A PyTorch template project"
authors = [
    { name = "Your Name", email = "your.email@example.com" }  # Replace with your info
]
requires-python = ">=3.7"
dependencies = [
    "PyYAML>=6.0",
    "torch>=1.13.1",
    "torchvision>=0.14.1",
    "torchinfo>=1.7.2",
    "wandb>=0.13.11",
    "scipy>=1.10.1",
    "tqdm>=4.65.0",
    "matplotlib>=3.7.1",
    "opencv-python>=4.5.5"  # Add OpenCV here
]

[build-system]
requires = ["setuptools", "wheel", "setuptools_scm"]
build-backend = "setuptools.build_meta"
